+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD (unreleased)
 
+- Feature (`ngx-column`): Added a new column component
+
 ## 49.2.1 (2025-06-18)
 
 - Fix (`ngx-calendar`): Improved Firefox compatibility by replacing outline: auto with outline: none